Create replace_stringtie_ID_merge_counts.py
Python3 script for replacing the MSTRG IDs introduced by StringTie with the reference IDs/ names and combining them with an existing count matrix. This script can be used, for example, after running nf-core/nanoseq or rnaseq etc. to obtain a single table containing the reference IDs and the count data. The script can also be easily adjusted for processing any gtf/ gff file and to split and extract the information of interest from the features column.